client-services: merge service-context into service-host
Address review feedback: the layer composition in service-context.ts was small
enough to fold into service-host.ts. Move ServiceContextLayer and the store/echo
/replicator/feed-syncer layers there, and relocate the two symbols the lifecycle
module depends on (StorageMigrationService, ServiceContextComponents) into
service-lifecycle.ts to avoid a value-import cycle (service-host imports
service-lifecycle). service-context.ts is deleted.
