+++ packages/sdk/client-services/src/packlets/services/index.ts
@@ -5,7 +5,6 @@
 export * from './client-services-layer';
 export * from './cross-device-space-synchronizer';
 export * from './feed-syncer';
-export * from './service-context';
 export * from './service-lifecycle';
 export * from './service-host';
 export * from './service-registry';
